format: store lengths with scale 1, not 4
This is needed for #7

+++ format/src/main/java/com/cldellow/manu/format/FieldEncoder.java
@@ -9,7 +9,7 @@ public interface FieldEncoder {
     boolean isVariableLength();
     int getLength();
 
-    void encode(int[] data, int[] encoded, IntWrapper encodedLength);
+    void encode(int[] data, byte[] encoded, IntWrapper encodedLength);
 
-    void decode(int[] encoded, int encodedLength, int[] data, IntWrapper dataLength);
+    void decode(byte[] encoded, int encodedLength, int[] data, IntWrapper dataLength);
 }
